cam_bot/requirements.txt
numpy==1.20.3
opencv-python==4.5.2.52
pyserial==3.5
urllib3==1.26.5
